requirements.txt
aioredis
alabaster
anyio
asgiref
async-timeout
Babel
base58
bitarray
bitstring
boto3
botocore
certifi
chardet
charset-normalizer
cleo
click
crashtest
Deprecated
distlib
dnspython
docutils
ECPy
fastapi
filelock
h11
hiredis
httpcore
httpx
idna
imagesize
importlib-metadata
Jinja2
jmespath
json2html
MarkupSafe
motor
nose
numpy
packaging
pandas
pdfkit
pipenv
platformdirs
pptree
psycopg2
pydantic>=v1.10.0a1
Pygments
pylev
pymongo
pyparsing
python-dateutil
python-dotenv
python-multipart
python-ulid
pytz
redis
redis-om
requests
rfc3986
s3transfer
six
sniffio
snowballstemmer
Sphinx
starlette
tabulate
types-redis
types-six
typing-extensions
urllib3
uvicorn
websocket-client
websockets
wheel
wkhtmltopdf
wrapt
xrpl-py>=1.5.0
xumm-sdk-py
zipp
gunicorn
opentelemetry-distro
opentelemetry-exporter-otlp
